feat: Descriptive error messages
Returning the HTTP error response code isn't as informative
as a message that describes what went wrong.

This commit will get the GotError block to return a description
of what went wrong.

I've tried to implement this in a way that's consistent with
the existing code, such as the approach to parsing the error
JSON payload.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>