feat: reuse existing image blocks for imgtfjs projects
I really didn't like the idea of introducing new separate blocks
to the extension just for the on-device image projects, so I've
tried to overload the existing image blocks so they can do both
server-based and on-device based approaches.

To do this, I needed to introduce the idea of the extension
knowing what type of ML project the student is doing (to tell
the difference between images and imgtfjs) so while I did that
I added some usability improvements for text and numbers, too.

It means students will get a more helpful error message if they
use blocks don't match with the type of API key they are using
or if they try to use invalid API keys.

I've also done it in a way that I think (hope?) will enable
support for sound projects in the future.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>