chore: remove dead code for the now-removed "images" project type
The "images" project type used the IBM Visual Recognition service
which is no longer available. This commit is a tiny bit of tidying
up to remove the code that handled that type. (It's not super
essential, as at the moment that code is just never called).

I've also updated the version of TensorFlow used in the extension
as there are a bunch of fixes and performance improvements that
went into the version since we last built this.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>